// src/app/core/models/api/response/user-response.model.ts
export class UserResponse {
    BUDescription: string;
    CreatedDate: string;
    Email: string;
    FirstName: string;
    FullName: string;
    ID: number;
    IsDeleted: boolean;
    LastLoginDate: string;
    LastName: string;
    MeridiumSecret: string;
    MeridiumUserName: string;
    Type: number;
    UpdatedBy: number;
    Updateddate: string;
    acknowledgedDisclaimer: string;
    acknowledged_update: string;
}